+++ Source/ACE/Entity/BroadcastAction.cs
@@ -8,6 +8,8 @@
 
         LocalChat,
 
-        Sound
+        PlaySound,
+
+        PlayParticleEffect
     }
 }